services/preprocessing_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv
pydantic
scikit-learn
pytest
httpx